requirements.txt
python-dotenv==1.0.0
openai==1.12.0
mistralai==1.0.0
numpy==1.26.3
pandas==2.1.4
sentence-transformers==2.3.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
PyYAML==6.0.1
tqdm==4.66.1